javascript updates
Removed the 'return true' statement from the if statement because it
can only be used in a function. This was breaking the tools on safari.
Also added in a check to remove any urls that had a # at the end. This
was caused when someone used the "Go" Button and then pressed enter on
the input (or added it in directly). It caused laytout issues with the
iframes.